fix(billing): remediation layout, split modules, tests, validation
Admin web (PR 1477 remediation):
- Invoice Operations: Issue, Email dialog, Void dialog; remove editor void/confirm duplicate
- Draft invoice uses AdminEditorCard actions; allocate/refund in separate editor cards
- Selected invoice Card directly under invoice table; currency filter toolbar; AbortSignal on loads
- Clear banners on selection change; seed email/payment fields without clobbering user input
- Rename formatTruncatedId; stricter parseLineTotalsOverridesJson with structured errors
- ClientInvoicesPanel RTL tests; finance-page assertion update

Backend:
- Split admin_billing into admin_billing_common, payments, invoice_queries/serializers,
  invoices (mutations), allocations, export; router stays under 500 lines each
- ValidationError for bad UUID/Decimal in refund, allocation, draft enrollments
- Email format check for invoice email API
- Document tenant-wide list in OpenAPI; lambdas.md notes module split

Tests: expand test_admin_billing for list filters, cursor, currency length, email, amount
