+++ apps/admin_web/tests/components/admin/finance/finance-page.test.tsx
@@ -104,7 +104,7 @@ describe('FinancePage', () => {
 
     await user.click(screen.getByRole('button', { name: 'Client Invoices' }));
     expect(screen.getByRole('heading', { name: 'Client Invoices' })).toBeInTheDocument();
-    expect(screen.getByRole('heading', { name: 'Client billing actions' })).toBeInTheDocument();
+    expect(screen.getByRole('heading', { name: 'Create draft invoice from enrollments' })).toBeInTheDocument();
     expect(window.location.search).toBe('?tab=client-invoices');
 
     await user.click(screen.getByRole('button', { name: 'Expenses' }));
