+++ docs/api/admin.yaml
@@ -2939,6 +2939,8 @@ paths:
       description: >
         Cursor-paginated list ordered by `created_at` descending, then `id` descending.
         Pass `next_cursor` from the previous response as `cursor` for the next page.
+        Returns all invoices for the tenant; **admin authorization is required** (API Gateway admin group).
+        Filter by optional `currency` (three-letter ISO code) and `status`.
       security:
         - AdminBearerAuth: []
       parameters:
